pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nanobots"
version = "0.1.0"
description = "Fire-and-forget micro-agents for AI systems. Spawn, execute, self-destruct."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Ava Shakil", email = "ava@artifactvirtual.com" },
    { name = "Ali Shakil", email = "ali@artifactvirtual.com" },
]
keywords = ["ai", "agents", "micro-agents", "nanobots", "automation", "devops", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Monitoring",
    "Topic :: Security",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
nanobot = "nanobots.cli:main"

[project.urls]
Homepage = "https://github.com/ArtifactVirtual/nanobots"
Documentation = "https://github.com/ArtifactVirtual/nanobots#readme"
Repository = "https://github.com/ArtifactVirtual/nanobots"
Issues = "https://github.com/ArtifactVirtual/nanobots/issues"

[tool.hatch.build.targets.sdist]
include = ["/src/nanobots"]

[tool.hatch.build.targets.wheel]
packages = ["src/nanobots"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
